* Made sure that when hitting the iPad's hide keyboard button on the "search" view, it actually dismissed the autocomplete popover as well. Otherwise we have this issue where the auto popover has a white peek through at the bottom and does not coincide with the the autocomplete view being shown. So this is dismissed when the hide button is tapped * Added constraints to the web toolbar to follow the width of the screen (saves from having the half border look when tapping into a story, the issue was observed on an iPad) * Made changes to the AutoHider to reject any offsets that are in the negative coordinates, think pull to refresh, when this happens the toolbar and nav bar gets funky and toggles between compact (hidden) and expanded (shown) states, nasty. * Made changes to the AutoHide to ensure that the "auto hide" methods do not get called when the offset is greater than the content. Eg scrolling up and then pulling the content up from the bottom used to cause a spasm. * Replaced UIWebView with WKWebView to support auto-resizing of image only content (Reddit, Imgurl) * Made some changes to enable WKWebView support cached / file content. * Fixed an issue with opening _blank links in WKWebView
There was an issue when the user is in the Search module then tapping on a recent or even on an item and the DDG orange would be shown on the bottom of the web view:
* Removed the bottom constraint that moved the web view up
* Updated the keyboard notifications to listen to didShow now willShow
* Ensured that we are now using contentInsets of the Autocomplete tableView to give more space. It looks better as there's no peek through of the orange DDG background

Updated the UIWebView to WKWebView for two reasons, one is to support auto single image scaling. There were times when we were getting single images coming it at 3000+ px and they were not being resized in the screen mostly due to now being wrapped in the appropriate HTML/CSS. So instead of doing our own checks, UIWebView has been replaced by WKWebView in iOS8 and above. This gives us the auto fit support but also improves performance as it uses Javascript's Nitro engine and it's more efficient.