* Fix for slow loading of the search bar * Running DDG though Xcode Instruments using Time Profile showed that in setShowsBangButton, layoutIfNeeded was causing a significant delay, removed this and tested to ensure that there no adverse effects, search bar is quicker to open
* Fixes for the layout of Recent's to not render *
After further testing, we found that when a History entity was added into the MOC the fetchResultsController's controllerDidChangeContent was called but there was an error updating the CollectionView, this error would free the layout. By setting the fetchedResultsController in viewWIllAppear this sets it's delegate and by setting the delegate to nil in viewWillDisappear this ensures that the View controller that's visible is the only one to receive the updates.

* Further fixes to UICollectionView updates *
When there's a NSFetchedResultsChangeUpdate in the objectChanges array these values are collated and then executed upon completion, this is due to issues with calling reloadItemsAtIndexPaths inside a collectionVIews performBatchUpdates: block. PerformBatchUpdates needs to have insert, updates or removal with the pre change and post change values equaling changes (insert, removal etc).

* Search bar url loading fixes *
Sometimes the url and Search DuckDuckGo label would be displayed at the same time, when loading a Story. In DDGStory's writeHTMLString:completion: method, upon success it used to call the completion block on a thread that was not the main thread, thus there were issues with the system saying that we are calling updates to the auto layout system from the background thread. Placing the call inside of the GCD get_main_queue block ensures that it get's called on the main thread. We have seen a decrease in loading times of stories from this.